% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCAGEDdataR-package.R
\docType{package}
\name{NCAGEDdataR-package}
\alias{NCAGEDdataR}
\alias{NCAGEDdataR-package}
\title{NCAGEDdataR: Fácil download e extração dos dados do Novo CAGED}
\description{
\if{html}{\figure{NCAGEDdataR_logo1.png}{options: style='float: right' alt='logo' width='120'}}

O 'NCAGEDdataR' é um pacote R que permite aos usuários acessar e analisar conjuntos de dados do NOVO CAGED (portal PDET) utilizando uma arquitetura de alta performance, estruturando as informações em datasets consolidados para consultas rápidas e de baixo consumo de memória.

Saiba mais sobre a documentação e atualizações em \url{https://natanaelsl.com.br/project/caged_pkg}.
}
\details{
O pacote automatiza o ciclo de vida do dado: desde a captura via API até a
persistência otimizada em formato Apache Parquet.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Natanaelsl/NCAGEDdataR}
  \item \url{https://natanaelsl.com.br/project/caged_pkg/}
  \item Report bugs at \url{https://github.com/Natanaelsl/NCAGEDdataR/issues}
}

}
\author{
\strong{Maintainer}: Natanael Soares Leite \email{natanaelsoaresleite@gmail.com} (\href{https://orcid.org/0000-0003-1845-7095}{ORCID})

}
\keyword{internal}
