\docType{package}
\name{NCAGEDdataR}
\alias{NCAGEDdataR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
NCAGEDdataR: Fácil download e extração dos dados do Novo CAGED
}
\description{
\if{html}{\figure{NCAGEDdataR_logo1.png}{options: style='float: right' alt='NCAGEDdataR_logo2' width='120'}}

O 'NCAGEDdataR' é um pacote R que permite aos usuários acessar facilmente conjuntos de dados do NOVO CAGED. Saiba mais sobre o 'NCAGEDdataR' em \url{https://natanaelsl.com.br/project/caged_pkg}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://natanaelsl.com.br/project/caged_pkg/}
  \item \url{https://github.com/Natanaelsl/NCAGEDdataR}
  \item Report bugs at \url{https://github.com/Natanaelsl/NCAGEDdataR/issues}
}
}

\author{
\itemize{
  \item Natanael Soares Leite (\href{https://orcid.org/0000-0003-1845-7095}{ORCID})
}
%\email{hadley@gmail.com}
}
\keyword{internal}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
