% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdata.R
\name{NCdata}
\alias{NCdata}
\title{NCAGEDdataR: Extra????o de dados do NOVO CAGED}
\usage{
NCdata(year = NULL, month = NULL)
}
\arguments{
\item{year}{valor num??rico}

\item{month}{string para o m??s}
}
\value{
Uma lista de data.frame
}
\description{
'
}
