% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdata.R
\name{NCdata}
\alias{NCdata}
\title{NCdata: Extração de dados do NOVO CAGED}
\usage{
NCdata(year = NULL, month = NULL)
}
\arguments{
\item{year}{Ano (valor numérico)}

\item{month}{Mês (string)}
}
\value{
Uma lista de data.frames para cada uma das abas da planilha do NOVO CAGED.
}
\description{
Realiza o extração dos dados, para o ano e mês especificados, retornando como lista de data.frame cada uma das abas do aquivo .xlsx dísponibilizado no site do Ministério do Trabalho - PDET.
}
\examples{
# Extraindo dados de Janeiro/2023
NCdata(2023, "Janeiro")

}
