% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdata.R
\name{NCdata}
\alias{NCdata}
\title{NCAGEDdataR: Extração de dados do NOVO CAGED}
\usage{
NCdata(ano = NULL, mês = NULL)
}
\arguments{
\item{ano}{valor numérico}

\item{mês}{string para o mês}
}
\description{
NCAGEDdataR: Extração de dados do NOVO CAGED
}
