% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdata.R
\name{NCdata}
\alias{NCdata}
\title{NCdata: Extração de dados do NOVO CAGED}
\usage{
NCdata(year = NULL, month = NULL)
}
\arguments{
\item{year}{valor numérico}

\item{month}{string para o mês}
}
\value{
Uma lista de data.frame para cada um dos 'sheet' da planilha do NOVO CAGED.
}
\description{
Realiza o extração dos dados, para o ano e mês especificados, retornando como lista de data.frame cada uma das abas do aquivo .xlsx dísponibilizado no site do Ministério do Trabalho - PDET.
}
\examples{
NCdata(2023, "Janeiro")

}
