% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdata.R
\name{NCdata}
\alias{NCdata}
\title{NCAGEDdataR: Extração de dados do NOVO CAGED}
\usage{
NCdata(year = NULL, month = NULL)
}
\arguments{
\item{year}{valor numérico}

\item{month}{string para o mês}
}
\value{
Uma lista de data.frame
}
\description{
'
}
\examples{
NCdata(2023, "Janeiro")


}
