% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdown.R
\name{NCdown}
\alias{NCdown}
\title{NCdown: Download dos dados do NOVO CAGED}
\usage{
NCdown(year = NULL, month = NULL, path = NULL)
}
\arguments{
\item{year}{valor numérico}

\item{month}{string para o mês}

\item{path}{diretório para salvar os dados baixados}
}
\value{
Salva a planilha do PDET com os dados do NOVO CAGED no diretório informado.
}
\description{
Realiza o download do aquivo .xlsx no site do Ministério do Trabalho - PDET com os dados do NOVO CAGED para o ano e mês especificados.
}
